added post for appointment, credentials and put fpr donation form and pacient by donorcode

+++ OneBlood/src/main/java/ac/OneBlood/Repository/PersonalInformationRepository.java
@@ -4,6 +4,8 @@ import ac.OneBlood.Model.PersonalInformation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface PersonalInformationRepository extends JpaRepository<PersonalInformation, Integer> {
+public interface PersonalInformationRepository extends JpaRepository<PersonalInformation, BigInteger> {
 }
